+++ src/components/range/range.ts
@@ -4,7 +4,7 @@ import { classMap } from 'lit/directives/class-map.js';
 import { ifDefined } from 'lit/directives/if-defined.js';
 import { live } from 'lit/directives/live.js';
 import { emit } from '~/internal/event';
-import { FormSubmitController } from '~/internal/form-control';
+import { FormSubmitController } from '~/internal/form';
 import { HasSlotController } from '~/internal/slot';
 import { watch } from '~/internal/watch';
 import styles from './range.styles';
@@ -20,6 +20,10 @@ import styles from './range.styles';
  * @event sl-blur - Emitted when the control loses focus.
  * @event sl-focus - Emitted when the control gains focus.
  *
+ * @csspart form-control - The form control that wraps the label, input, and help-text.
+ * @csspart form-control-label - The label's wrapper.
+ * @csspart form-control-input - The range's wrapper.
+ * @csspart form-control-help-text - The help text's wrapper.
  * @csspart base - The component's internal wrapper.
  * @csspart input - The native range input.
  * @csspart tooltip - The range tooltip.
@@ -213,11 +217,16 @@ export default class SlRange extends LitElement {
           'form-control--has-help-text': hasHelpText
         })}
       >
-        <label part="label" class="form-control__label" for="input" aria-hidden=${hasLabel ? 'false' : 'true'}>
+        <label
+          part="form-control-label"
+          class="form-control__label"
+          for="input"
+          aria-hidden=${hasLabel ? 'false' : 'true'}
+        >
           <slot name="label">${this.label}</slot>
         </label>
 
-        <div class="form-control__input">
+        <div part="form-control-input" class="form-control-input">
           <div
             part="base"
             class=${classMap({
@@ -260,7 +269,7 @@ export default class SlRange extends LitElement {
         </div>
 
         <div
-          part="help-text"
+          part="form-control-help-text"
           id="help-text"
           class="form-control__help-text"
           aria-hidden=${hasHelpText ? 'false' : 'true'}